refactor: moved SyncedJSON to json package
